Make it also understand and handle extended BetterString MCC Features. Define new MUIA_String_#?/MUIM_String_#? attrs/methods based on the MUIA_BetterString_#?/MUIM_BetterString_#? ones. The used IDs are exactly the same ones as used by BetterString MCC. This way we can later make a dummy BetterString Class as subclass of String Class and everything is done by the string class. Or as alternative add some "BetterString.mcc" to "String.mcc" class name mapping to muimaster.library.
Fixed also some compiler warnings and a crash which happened when a
char was supposed to be added while the cursor was at the beginning and
there was no more space in the buffer for that space. Crashed in MUIM_Draw
addchar draw optimization because of "bufferpos - 1"'s where bufferpos was 0.
BufferPos is normally never 0 after a AddChar, but if AddChar failed it may
stay 0. That's where the problem was.


git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@22088 fb15a70f-31f2-0310-bbcc-cdcc74a49acc