The shape of the Window title was prevoiusly rendered in the DrawWinBorder method, this is now moved to the WindowShape method.
There is one problem left, to be exact all gadgets must be installed before the DefaultShapeHook is called,
this isn´t yet. For decorations that meant that the Titlebar can not be exactly created because Systemgadgets
was installed after the WindowShape method is called. But Decoration needs all gadgets to layout the Titlebar
and the Shape.

The problem is that when using the Mask of the Titlebar the Shape is wrong after the Window appears.

So solve this, the Gadgets must be created before the DefaultShape Hook is called, or the Windowframe must be refreshed later.

I´ve no idea how to solve this, i hope Georg has any ideas.

Darius

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@26034 fb15a70f-31f2-0310-bbcc-cdcc74a49acc