Removed DHCP column in list. Instead print "DCHP" in the IP and Mask column.
Added column "Active" which is related to the "UP" argument in the interface
configuration file.

The DHCP Cycle gadget for Gate and DNS.

Don't store DNS entries when DHCP is active (as it was the case before my fixes.)

Don't store Gate entry when DCHP is active (has to be tested)
(This is probably a flaw in our TCP stack. On my Unix box the gateway belongs
to the interface spec.)

Copy Kal's image to Image: and use it in the interface window.

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@32209 fb15a70f-31f2-0310-bbcc-cdcc74a49acc