arch/m68k-amiga/boot: Split ROM
The ROM is now split. The 'rom' image will contain everything
needed to boot from disk, and the 'ext' image will contain
everything that is not needed for that task. Ideally, the
libraries and resources on the 'ext' disk could be on the AFS
filesystem.

For now, intuition, graphics, and layers live on 'ext'.

0xf80000-0xffffff: a-m-a-rom.bin
0xe00000-0xe7ffff: a-m-a-ext.bin

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@35257 fb15a70f-31f2-0310-bbcc-cdcc74a49acc