arch/m68k-amiga: Add kernel-link-amiga-m68k-ram target (WIP)
The 'distfiles/aros-amiga-m68k-ram.{elf,bin}' files generated
are designed to be loaded at address 0xc00000 on machines
that have that RAM, to allow testing of the AROS m68k ROM
when people don't have the ability to use ROm mapping hardware,
such as the ACA 1230/56.

Just load the image at 0xc00000-0xcfffff, and jump to 0xc00000.

This is just the (untested) infrastructure, will probably need
more debugging.

Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@36483 fb15a70f-31f2-0310-bbcc-cdcc74a49acc