arosc.library/librom.a: %f/%g vsprintf support is only in arosc.library, not librom.a
Needed for m68k kernel builds.

The librom.a library now has the old-style 'no %f/%g' printf family,
arosc.library has the floating point stuff.

So, if you want %f, use 'linklibs="arosc"', and *not* linklibs="rom"

It may be prudent to have the 'rom' version print out '???' or
something if someone accidentally links with it and uses a '%f' format.

We'll add that in a later revision, if needed.

Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@37358 fb15a70f-31f2-0310-bbcc-cdcc74a49acc