m68k/AROSBootstrap: Proof of concept, work in progress
This is a proof of concept m68k AROS bootstrap, that is designed
to run on AmigaOS 1.3 or better. I have some more debugging
to do, but this will:

* Decompress and load a 'aros.elf.gz' file into MEMF_KICK RAM
  performing all necessary relocations
* Bind its memory to the SysBase->KickMemPtr list
* Add the initial vector of the ROM to the old SysBase's
  ColdCapture vector
* Restart into the AROS ROM.

At this time, AROS quickly dies after Expansion.library
due to a freelist corruption, but I should be able to
debug that soon.

Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@37497 fb15a70f-31f2-0310-bbcc-cdcc74a49acc