WindowDecorClass - cache assembled prop gadgets images and reuse them instead of reassemly
Assembling a complete image of prop gadget requires bitmap allocation and
several WritePixelAlpha calls. This is slow and make moving windows on top
of other windows slow as well (because move of window reveals underlying
windows and causes a full gadget redraw on them). By caching the bitmap
and detecting if the gadget has changed, the slow assembly process is remove
and redrawing windows is faster. Also bitmaps are not constantly allocated/
deallocated which removes additional performance hit.

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@37819 fb15a70f-31f2-0310-bbcc-cdcc74a49acc