Added support for library libbases to be stored as an offset in the current relbase pointer.
A global variable libbasename_offset indicates at which offset in the
current the other libbase is stored. Support macro's are available in
symbolsets.h.

Support is also added in genmodule so that for all modules compiled with
%build_module generates a liblibname_rel.a file that will call the
function using this offset in the current libbase.

If you want to place a libbase of another library in the library of your
library you have to do the following things:
- Put a field in your libbase
- Initialize the libbasename_offset variable properly
- Link the library with liblibase_rel.a evt. using uslibs=libbase_rel

This patch is co-developed with Jason McMullan.

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@41080 fb15a70f-31f2-0310-bbcc-cdcc74a49acc