tools/genmodule: Added .unusedlibbase option
This option allows to indicate that the libbase does not need to be accessible inside the library when entering through this function.
Using this function can avoid some overhead internally when calling such functions.
One needs to be careful with using this option as the .unusedlibbase function may not call other functions that want to access the libbase.

Added test case in test/library for the option.

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@45652 fb15a70f-31f2-0310-bbcc-cdcc74a49acc