icon.library: trim garbage that can be present at end of file
When (un)snapshoting, icon is saved with ICONPUTA_OnlyUpdatePosition. This
causes the file to be opened in MODE_OLDFILE. Since X/Y coordinates are
not saved when they are equal NO_ICON_POSITION, the data is smaller and
there is garbage left at the end of the file. Not trimming the file was
causing the icon to be damaged after unsnapshoting.

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@47666 fb15a70f-31f2-0310-bbcc-cdcc74a49acc