posixc.library: use directly Input() and Output() of process
Previous code duplicated the handles when creating C std in and out
streams. This meant that operations on C std streams were not reflected
in the process handled as well as that duplicating reset the file pointer
to the beginning. This caused data to be overwritten when the intention
was to append to output. Output() was opened and seek to the end, but
the copied out stream was pointed at the beginning.

git-svn-id: https://svn.aros.org/svn/aros/trunk/AROS@48996 fb15a70f-31f2-0310-bbcc-cdcc74a49acc